@@ -10,33 +10,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.2.0",
-    "@angular/common": "~13.2.0",
-    "@angular/compiler": "~13.2.0",
-    "@angular/core": "~13.2.0",
-    "@angular/forms": "~13.2.0",
-    "@angular/localize": "~13.2.0",
-    "@angular/platform-browser": "~13.2.0",
-    "@angular/platform-browser-dynamic": "~13.2.0",
-    "@angular/router": "~13.2.0",
-    "@ng-bootstrap/ng-bootstrap": "^12.0.0",
+    "@angular/animations": "^16.1.3",
+    "@angular/common": "^16.1.3",
+    "@angular/compiler": "^16.1.3",
+    "@angular/core": "^16.1.3",
+    "@angular/forms": "^16.1.3",
+    "@angular/localize": "^16.1.3",
+    "@angular/platform-browser": "^16.1.3",
+    "@angular/platform-browser-dynamic": "^16.1.3",
+    "@angular/router": "^16.1.3",
+    "@ng-bootstrap/ng-bootstrap": "^15.0.1",
     "@ngneat/until-destroy": "^9.2.2",
-    "@ngrx/effects": "^13.1.0",
-    "@ngrx/store": "^13.1.0",
-    "@ngrx/store-devtools": "^13.1.0",
-    "@popperjs/core": "^2.10.2",
+    "@ngrx/effects": "^15.4.0",
+    "@ngrx/store": "^15.4.0",
+    "@ngrx/store-devtools": "^15.4.0",
+    "@popperjs/core": "^2.11.8",
     "bootstrap": "^5.1.3",
     "bootstrap-icons": "^1.10.3",
     "chart.js": "^3.9.1",
     "ng2-charts": "^3.1.2",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.11.4"
+    "zone.js": "~0.13.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.5",
-    "@angular/cli": "~13.2.5",
-    "@angular/compiler-cli": "~13.2.0",
+    "@angular-devkit/build-angular": "^16.1.3",
+    "@angular/cli": "^16.1.3",
+    "@angular/compiler-cli": "^16.1.3",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
     "jasmine-core": "~4.0.0",
@@ -45,6 +45,6 @@
     "karma-coverage": "~2.1.0",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.5.2"
+    "typescript": "~4.9.5"
   }
 }
