@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Router, CanActivate } from '@angular/router';
+import { Router } from '@angular/router';
 import { AccountService } from '../account/services/account.service';
 import { UntilDestroy } from '@ngneat/until-destroy';
 import { Observable, map } from 'rxjs';
@@ -8,7 +8,7 @@ import { Observable, map } from 'rxjs';
   providedIn: 'root',
 })
 @UntilDestroy()
-export class AuthGuardService implements CanActivate {
+export class AuthGuardService  {
   constructor(public accountService: AccountService, public router: Router) {}
 
   canActivate(): Observable<boolean> {
